Refactor to Client DNS module
