+++ elixir/apps/domain/lib/domain/config/configuration.ex
@@ -1,13 +1,9 @@
 defmodule Domain.Config.Configuration do
   use Domain, :schema
-  alias Domain.Config.Logo
+  alias Domain.Config.{ClientsUpstreamDNS, Logo}
 
   schema "configurations" do
-    embeds_many :clients_upstream_dns, ClientsUpstreamDNS, on_replace: :delete, primary_key: false do
-      field :type, :string
-      field :address, :string
-    end
-
+    embeds_many :clients_upstream_dns, ClientsUpstreamDNS, on_replace: :delete
     embeds_one :logo, Logo, on_replace: :delete
 
     belongs_to :account, Domain.Accounts.Account
